requirements.txt
aiofiles==0.3.2
aiohttp==3.1.3
async-timeout==2.0.1
attrs==17.4.0
chardet==3.0.4
httptools==0.0.10
idna==2.6
idna-ssl==1.0.1
Jinja2==2.10
MarkupSafe==1.0
motor==1.2.1
multidict==4.2.0
PyJWT==1.6.1
pymongo==3.6.1
python-rapidjson==0.4.3
sanic==0.7.0
ujson==1.35
uvloop==0.9.1
websockets==4.0.1
yarl==1.2.2
